Core/Entities: Final batch of removing implicit conversions of ObjectGuid to uint64
(cherry picked from commit 68f45e87345ff85341f835f0372f2a62b89f24d9)

+++ src/server/game/Entities/Item/Item.cpp
@@ -1019,7 +1019,7 @@ bool Item::IsLimitedToAnotherMapOrZone(uint32 cur_mapId, uint32 cur_zoneId) cons
 void Item::SendUpdateSockets()
 {
     WorldPacket data(SMSG_SOCKET_GEMS_RESULT, 8+4+4+4+4);
-    data << uint64(GetGUID());
+    data << GetGUID();
     for (uint32 i = SOCK_ENCHANTMENT_SLOT; i <= BONUS_ENCHANTMENT_SLOT; ++i)
         data << uint32(GetEnchantmentId(EnchantmentSlot(i)));
 
@@ -1036,7 +1036,7 @@ void Item::SendTimeUpdate(Player* owner)
         return;
 
     WorldPacket data(SMSG_ITEM_TIME_UPDATE, (8+4));
-    data << uint64(GetGUID());
+    data << GetGUID();
     data << uint32(duration);
     owner->SendDirectMessage(&data);
 }
